Split out state.lua
